requirements.txt
# PDFToolKit - Python Bağımlılıkları

# GUI
PyQt6>=6.6

# PDF İşleme
pypdf>=4.0
PyMuPDF>=1.23

# OCR
pytesseract>=0.3

# Word Dönüştürme
python-docx>=1.1
pdf2docx>=0.5

# Excel Dönüştürme
openpyxl>=3.1
fpdf2>=2.7

# Resim İşleme
Pillow>=10.0

# Dosya Silme (Geri Dönüşüm Kutusu)
send2trash>=1.8

# Windows COM Otomasyonu (Word/Excel → PDF birebir dönüşüm)
pywin32>=306; sys_platform == "win32"

# Ortam Yönetimi
python-dotenv>=1.0

# Test
pytest>=7.0
